pyproject.toml
[project]
name = "retail-forecasting"
version = "0.1.0"
description = "AMLC community talk: retail sales forecasting with cyclicality analysis"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "nbconvert>=7.17.0",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "papermill>=2.7.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "statsmodels>=0.14.6",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "papermill>=2.7.0",
    "pytest>=9.0.2",
]
